+++ packages/vitest-config-custom/index.ts
@@ -1,4 +1,3 @@
-// @ts-ignore
 const sharedConfig = require("./vitest.shared.config.ts");
 
 module.exports = {